island(#6): move settings store + RestComms into island (the mutually-coupled v1-comms pair)
The biggest single migration — settings (964 LOC) + RestComms (731 LOC) move
together because they're mutually coupled (settings calls restComms.writeUserSettings
for cloud-sync; RestComms reads the settings store for the backend URL). Moving the
pair into island resolves BOTH the circular dependency AND the early-auth timing trap
(RestComms reads the backend URL at exchangeToken, before island's config/adapters are
wired — only the module-load-available store works there; now it's island's own store).

- settings.ts -> island/src/stores; RestComms.ts -> island/src/services (git mv,
  history preserved). Host @/stores/settings + @/services/RestComms are re-export shims
  (96 + 24 consumers unchanged). Exposed as toolkit.stores.settings.
- Import rewrites: settings -> island RestComms + island storage; RestComms -> island
  settings/connection stores, island GlobalEventEmitter, island types, BgTimer, and the
  internal btsdk surface via the relative build/_internal path (the
  @mentra/bluetooth-sdk-internal alias doesn't resolve in island's standalone build).
- GlobalEventEmitter moved into island (one shared instance + host shim); ported
  storage.loadSubKeys into island's MMKV util; dropped two dead imports (Platform/Device).
- jest: requireActual settings + RestComms (tests used the real host store before the
  move); moduleNameMapper maps the relative build/_internal path to the mockable source
  so requireActual doesn't load the real native btsdk; the mock exposes the REAL island
  GlobalEventEmitter so emit/listen share one instance.

v1-transitional: RestComms is deleted in place when v1 retires; the point is the host
no longer owns comms during the (multi-week) transition. SocketComms is a separate
4-6 week refactor and moves last.

mobile typecheck + island standalone build + full jest (281 tests) all green.
