+++ mobile/modules/island/src/services/RestComms.ts
@@ -1,0 +1,735 @@
+import axios, {AxiosInstance, AxiosRequestConfig} from "axios"
+import {AsyncResult, Result, result as Res} from "typesafe-ts"
+
+import type {AppletInterface} from "../types"
+// Internal btsdk surface (RestComms needs updateBluetoothSettings, not on the
+// public entry). Relative path — `@mentra/bluetooth-sdk-internal` is a mobile-app
+// tsconfig alias that doesn't resolve in island's standalone build. Same path
+// island's index.ts uses, one level deeper from services/.
+import BluetoothSdk, {type PhotoResponseEvent} from "../../../bluetooth-sdk/build/_internal"
+import {SETTINGS, useSettingsStore} from "../stores/settings"
+import {useConnectionStore} from "../stores/connection"
+import {WebSocketStatus} from "../stores/connection"
+import GlobalEventEmitter from "../utils/GlobalEventEmitter"
+import {BgTimer} from "../utils/timers"
+
+interface RequestConfig {
+  method: "GET" | "POST" | "DELETE"
+  endpoint: string
+  data?: any
+  params?: any
+  requiresAuth?: boolean
+}
+
+class RestComms {
+  private static instance: RestComms
+  private readonly TAG = "RestComms"
+  private coreToken: string | null = null
+  private axiosInstance: AxiosInstance
+
+  private constructor() {
+    this.axiosInstance = axios.create({
+      headers: {
+        "Content-Type": "application/json",
+      },
+    })
+  }
+
+  public static getInstance(): RestComms {
+    if (!RestComms.instance) {
+      RestComms.instance = new RestComms()
+    }
+    return RestComms.instance
+  }
+
+  // Token Management
+  public setCoreToken(token: string | null): void {
+    this.coreToken = token
+    const tokenLen = token?.length ?? 0
+    console.log(
+      `${this.TAG}: Core token ${token ? "set" : "cleared"} - Length: ${tokenLen} - First 20 chars: ${
+        token?.substring(0, 20) || "null"
+      }`,
+    )
+
+    // Sync to native DeviceStore (and persist to SharedPreferences in BluetoothSdkModule when bridge runs)
+    const value = token ?? ""
+    const updateResult = BluetoothSdk.updateBluetoothSettings({core_token: value})
+    if (updateResult != null && typeof (updateResult as Promise<void>).then === "function") {
+      ;(updateResult as Promise<void>).catch(() => {})
+    }
+
+    if (token) {
+      console.log(`${this.TAG}: Core token set, emitting CORE_TOKEN_SET event`)
+      GlobalEventEmitter.emit("CORE_TOKEN_SET")
+    }
+  }
+
+  public getCoreToken(): string | null {
+    return this.coreToken
+  }
+
+  // Helper Methods
+  private validateToken(): Result<void, Error> {
+    if (!this.coreToken) {
+      return Res.error(new Error("No core token available for authentication"))
+    }
+    return Res.ok(undefined)
+  }
+
+  private createAuthHeaders(): Record<string, string> {
+    return {
+      "Content-Type": "application/json",
+      "Authorization": `Bearer ${this.coreToken}`,
+    }
+  }
+
+  private makeRequest<T>(config: RequestConfig): AsyncResult<T, Error> {
+    const {method, endpoint, data, params, requiresAuth = true} = config
+
+    const baseUrl = useSettingsStore.getState().getRestUrl()
+    const url = `${baseUrl}${endpoint}`
+    // console.log(`REST: ${method}:${url}`)
+
+    const headers = requiresAuth ? this.createAuthHeaders() : {"Content-Type": "application/json"}
+
+    const axiosConfig: AxiosRequestConfig = {
+      method,
+      url,
+      headers,
+      data,
+      params,
+    }
+
+    return Res.try_async(async () => {
+      try {
+        const res = await this.axiosInstance.request<T>(axiosConfig)
+        return res.data
+      } catch (error) {
+        if (!this.isNoActiveSessionError(error)) {
+          throw error
+        }
+
+        // Cloud pod has no session for this user (we reconnected to a different
+        // pod, or the prior session was cleaned up). Trigger a WS reconnect,
+        // wait for it to land, then retry the request exactly once.
+        //
+        // Subscribe BEFORE emitting so we don't miss the DISCONNECTED → CONNECTED
+        // transition triggered by handleNoActiveSession → reconnectNow.
+        const waitPromise = this.waitForNextConnected(8_000)
+        GlobalEventEmitter.emit("NO_ACTIVE_SESSION")
+        try {
+          await waitPromise
+        } catch (waitErr) {
+          console.log(`${this.TAG}: Retry skipped — WS didn't reconnect in time:`, waitErr)
+          throw error
+        }
+
+        const retryHeaders = requiresAuth ? this.createAuthHeaders() : {"Content-Type": "application/json"}
+        const retryRes = await this.axiosInstance.request<T>({...axiosConfig, headers: retryHeaders})
+        return retryRes.data
+      }
+    })
+  }
+
+  /**
+   * Resolves on the NEXT CONNECTED transition of the WS (or rejects after
+   * timeoutMs). Does NOT short-circuit when already CONNECTED — callers
+   * invoke this after a 503 NO_ACTIVE_SESSION when we know the current
+   * connection is landing on the wrong pod; we need to wait for the
+   * post-reconnect CONNECTED event, not the current one.
+   *
+   * Uses the connection store directly rather than WebSocketManager to avoid
+   * a circular import (WebSocketManager → RestComms → WebSocketManager).
+   */
+  private waitForNextConnected(timeoutMs: number): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+      let settled = false
+      const timer = BgTimer.setTimeout(() => {
+        if (settled) return
+        settled = true
+        unsub()
+        reject(new Error(`waitForNextConnected timed out after ${timeoutMs}ms`))
+      }, timeoutMs)
+
+      let sawNonConnected = useConnectionStore.getState().status !== WebSocketStatus.CONNECTED
+      const unsub = useConnectionStore.subscribe((state) => {
+        if (settled) return
+        if (state.status !== WebSocketStatus.CONNECTED) {
+          sawNonConnected = true
+          return
+        }
+        // Only resolve on a CONNECTED transition that follows a non-CONNECTED
+        // state. This guarantees we waited for a real reconnect rather than
+        // resolving on the stale pre-reconnect CONNECTED state.
+        if (sawNonConnected) {
+          settled = true
+          BgTimer.clearTimeout(timer)
+          unsub()
+          resolve()
+        }
+      })
+    })
+  }
+
+  private isNoActiveSessionError(error: unknown): boolean {
+    if (!axios.isAxiosError(error)) {
+      return false
+    }
+
+    return error.response?.status === 503 && error.response?.data?.error === "NO_ACTIVE_SESSION"
+  }
+
+  private authenticatedRequest<T>(config: RequestConfig): AsyncResult<T, Error> {
+    let res = this.validateToken()
+    if (res.is_error()) {
+      return Res.error_async(res.error)
+    }
+    return this.makeRequest<T>({...config})
+  }
+
+  private unauthenticatedRequest<T>(config: RequestConfig): AsyncResult<T, Error> {
+    return this.makeRequest<T>({...config, requiresAuth: false})
+  }
+
+  // Public API Methods
+
+  public getMinimumClientVersion(): AsyncResult<{required: string; recommended: string}, Error> {
+    interface Response {
+      success: boolean
+      data: {required: string; recommended: string}
+    }
+    const config: RequestConfig = {
+      method: "GET",
+      endpoint: "/api/client/min-version",
+    }
+    const res = this.unauthenticatedRequest<Response>(config)
+    return res.map((response) => response.data)
+  }
+
+  public checkAppHealthStatus(packageName: string): AsyncResult<boolean, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/app-uptime/app-pkg-health-check",
+      data: {packageName},
+    }
+
+    interface Response {
+      success: boolean
+    }
+
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map((response) => response.success)
+  }
+
+  public retry<T>(fn: () => AsyncResult<T, Error>, attempts: number, delayMs: number = 0): AsyncResult<T, Error> {
+    return Res.try_async(async () => {
+      let lastError: Error | null = null
+
+      for (let i = 0; i < attempts; i++) {
+        const result: Result<T, Error> = await fn()
+        if (result.is_ok()) {
+          return result.value
+        }
+        lastError = result.error
+        if (i < attempts - 1 && delayMs > 0) {
+          await new Promise((resolve) => setTimeout(resolve, delayMs))
+        }
+      }
+      throw lastError
+    })
+  }
+
+  public getApplets(): AsyncResult<AppletInterface[], Error> {
+    interface Response {
+      success: boolean
+      data: AppletInterface[]
+    }
+    const config: RequestConfig = {
+      method: "GET",
+      endpoint: "/api/client/apps",
+    }
+    let res = this.authenticatedRequest<Response>(config)
+    let data = res.map((response) => response.data)
+    return data
+  }
+
+  public startApp(packageName: string): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: `/apps/${packageName}/start`,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public stopApp(packageName: string): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: `/apps/${packageName}/stop`,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public uninstallApp(packageName: string): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: `/api/apps/uninstall/${packageName}`,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  // App Settings
+  public getAppSettings(appName: string): AsyncResult<any, Error> {
+    const config: RequestConfig = {
+      method: "GET",
+      endpoint: `/appsettings/${appName}`,
+    }
+    const res = this.authenticatedRequest<any>(config)
+    return res
+  }
+
+  public updateAppSetting(appName: string, update: {key: string; value: any}): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: `/appsettings/${appName}`,
+      data: update,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map((response) => response.data)
+  }
+
+  public updateGlassesState(state: Record<string, any>): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/device/state",
+      data: state,
+    }
+    interface Response {
+      success: boolean
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public exchangeToken(token: string): AsyncResult<string, Error> {
+    const isChina: string = useSettingsStore.getState().getSetting(SETTINGS.china_deployment.key)
+
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/auth/exchange-token",
+      data: {
+        supabaseToken: !isChina ? token : undefined,
+        authingToken: isChina ? token : undefined,
+      },
+    }
+    interface Response {
+      coreToken: string
+    }
+    let res = this.makeRequest<Response>(config)
+    const coreTokenResult: AsyncResult<string, Error> = res.map((response) => response.coreToken)
+
+    // set the core token in the store:
+    return coreTokenResult.and_then((coreToken: string) => {
+      this.setCoreToken(coreToken)
+      return Res.ok(coreToken)
+    })
+  }
+
+  public generateWebviewToken(
+    packageName: string,
+    endpoint: string = "generate-webview-token",
+  ): AsyncResult<string, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: `/api/auth/${endpoint}`,
+      data: {packageName},
+    }
+    interface Response {
+      token: string
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map((response) => response.token)
+  }
+
+  public hashWithApiKey(stringToHash: string, packageName: string): AsyncResult<string, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/auth/hash-with-api-key",
+      data: {stringToHash, packageName},
+    }
+    interface Response {
+      hash: string
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map((response) => response.hash)
+  }
+
+  // Account Management
+  public requestAccountDeletion(): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/account/request-deletion",
+    }
+    interface Response {
+      success: boolean
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public confirmAccountDeletion(requestId: string, confirmationCode: string): AsyncResult<any, Error> {
+    const config: RequestConfig = {
+      method: "DELETE",
+      endpoint: "/api/account/confirm-deletion",
+      data: {requestId, confirmationCode},
+    }
+    interface Response {
+      success: boolean
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res
+  }
+
+  public getLivekitUrlAndToken(): AsyncResult<{url: string; token: string}, Error> {
+    const config: RequestConfig = {
+      method: "GET",
+      endpoint: "/api/client/livekit/token",
+    }
+    interface Response {
+      // url: string
+      // token: string
+      success: boolean
+      data: {url: string; token: string}
+    }
+    const res = this.authenticatedRequest<Response>(config)
+
+    // ;(async () => {
+    //   console.log("result@@@@@", await result)
+    //   // const response = await Res.value
+    //   // return {url: response.url, token: response.token}
+    // })()
+
+    return res.map((response) => response.data)
+  }
+
+  // User Feedback & Incidents
+
+  /**
+   * Create a new incident for a bug report.
+   * Returns incidentId for subsequent log/attachment uploads.
+   */
+  public createIncident(
+    feedback: object,
+    phoneState?: Record<string, unknown>,
+  ): AsyncResult<{success: boolean; incidentId: string}, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/incidents",
+      data: {
+        feedback,
+        ...(phoneState && {phoneState}),
+      },
+    }
+    interface Response {
+      success: boolean
+      incidentId: string
+    }
+    return this.authenticatedRequest<Response>(config)
+  }
+
+  /**
+   * Submit feedback (feature requests only).
+   * For bug reports, use createIncident instead.
+   */
+  public sendFeedback(
+    feedbackBody: string | object,
+    phoneState?: Record<string, unknown>,
+  ): AsyncResult<{success: boolean}, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/feedback",
+      data: {
+        feedback: feedbackBody,
+        ...(phoneState && {phoneState}),
+      },
+    }
+    interface Response {
+      success: boolean
+    }
+    return this.authenticatedRequest<Response>(config)
+  }
+
+  /**
+   * Upload phone logs to an incident.
+   * Called after createIncident returns an incidentId.
+   */
+  public uploadIncidentLogs(
+    incidentId: string,
+    logs: Array<{timestamp: number; level: string; message: string; source?: string}>,
+  ): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: `/api/incidents/${incidentId}/logs`,
+      data: {
+        source: "phone",
+        logs,
+      },
+    }
+    interface Response {
+      success: boolean
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  /**
+   * Upload screenshot attachments to an incident.
+   * Called after createIncident returns an incidentId.
+   */
+  public uploadIncidentAttachments(
+    incidentId: string,
+    images: Array<{uri: string; fileName?: string | null; mimeType?: string | null}>,
+  ): AsyncResult<{uploaded: number; errors: number}, Error> {
+    const uploadPromise = async (): Promise<Result<{uploaded: number; errors: number}, Error>> => {
+      try {
+        const coreToken = this.getCoreToken()
+        if (!coreToken) {
+          return Res.error(new Error("Not authenticated"))
+        }
+
+        const baseUrl = useSettingsStore.getState().getRestUrl()
+        const formData = new FormData()
+
+        for (const image of images) {
+          const filename = image.fileName || `screenshot-${Date.now()}.jpg`
+          const mimeType = image.mimeType || "image/jpeg"
+
+          // React Native FormData expects this format
+          formData.append("files", {
+            uri: image.uri,
+            name: filename,
+            type: mimeType,
+          } as unknown as Blob)
+        }
+
+        const response = await axios.post(`${baseUrl}/api/incidents/${incidentId}/attachments`, formData, {
+          headers: {
+            "Authorization": `Bearer ${coreToken}`,
+            "Content-Type": "multipart/form-data",
+          },
+          timeout: 60000, // 60 second timeout for uploads
+        })
+
+        const data = response.data as {
+          success: boolean
+          uploaded?: Array<{filename: string}>
+          errors?: Array<{filename: string; error: string}>
+        }
+
+        return Res.ok({
+          uploaded: data.uploaded?.length || 0,
+          errors: data.errors?.length || 0,
+        })
+      } catch (err) {
+        return Res.error(err instanceof Error ? err : new Error(String(err)))
+      }
+    }
+
+    return new AsyncResult(uploadPromise())
+  }
+
+  public writeUserSettings(settings: any): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/user/settings",
+      data: {settings},
+    }
+    interface Response {
+      success: boolean
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public loadUserSettings(): AsyncResult<any, Error> {
+    const config: RequestConfig = {
+      method: "GET",
+      endpoint: "/api/client/user/settings",
+    }
+    interface Response {
+      success: boolean
+      data: {settings: Record<string, any>}
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map((response) => response.data.settings)
+  }
+
+  // Error Reporting
+  public sendErrorReport(reportData: any): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/app/error-report",
+      data: reportData,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  // Calendar
+  public sendCalendarData(data: any): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/calendar",
+      data: data,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  // Location
+  public sendLocationData(data: any): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/location",
+      data: data,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  // Phone Notifications
+  public sendPhoneNotification(data: {
+    notificationId: string
+    app: string
+    title: string
+    content: string
+    priority: string
+    timestamp: number
+    packageName: string
+  }): AsyncResult<any, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/notifications",
+      data: data,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public sendPhoneNotificationDismissed(data: {
+    notificationId: string
+    notificationKey: string
+    packageName: string
+  }): AsyncResult<any, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/notifications/dismissed",
+      data: data,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public sendPhotoResponse(data: PhotoResponseEvent): AsyncResult<any, Error> {
+    const response =
+      data.state === "success"
+        ? {
+            type: data.type,
+            requestId: data.requestId,
+            photoUrl: data.uploadUrl,
+            timestamp: data.timestamp,
+            success: true,
+          }
+        : {
+            type: data.type,
+            requestId: data.requestId,
+            timestamp: data.timestamp,
+            success: false,
+            errorCode: data.errorCode,
+            errorMessage: data.errorMessage,
+          }
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/photo/response",
+      data: response,
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public goodbye(): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/goodbye",
+    }
+    interface Response {
+      success: boolean
+      data: any
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+
+  public configureAudioFormat(
+    format: string,
+    lc3Config: {
+      sampleRate: number
+      frameDurationMs: number
+      frameSizeBytes: number
+    },
+  ): AsyncResult<void, Error> {
+    const config: RequestConfig = {
+      method: "POST",
+      endpoint: "/api/client/audio/configure",
+      data: {format, lc3Config},
+    }
+    const res = this.authenticatedRequest<Response>(config)
+    return res.map(() => undefined)
+  }
+}
+
+const restComms = RestComms.getInstance()
+export default restComms
